Correct inherited feature policy when allow attribute is used.
This fixes an error in the inherited policy calculation where a frame
which was explicitly allowed to use a feature through the Feature-Policy
header would be blocked if the "allow" attribute was also present on
the iframe element.

The behaviour now matches the spec algorithm at
https://wicg.github.io/feature-policy/#define-inherited-policy-in-container

Tests have been added to catch this case.

Change-Id: I7a394202c615be3088fda738d7fc65f16bcadf34